Move _sqlite_adapters to connector file
